Rename file `errors.rb` => `error.rb`
