+++ lib/redmine_amzn_alb_authn/application_controller_patch.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative 'errors'
+require_relative 'error'
 require_relative 'oidc_data_decoder'
 
 module RedmineAmznAlbAuthn
